refactor: moved car functions to separate module
